change ip api to ipgeolocation.io

+++ internal/models/node.go
@@ -4,6 +4,6 @@ type Node struct {
 	IP         string
 	DomainName string
 	LatencyMs  int
-	Latitude   float32
-	Longitude  float32
+	Latitude   float64
+	Longitude  float64
 }
